Water Boss v0.10
The entities now gets summoned once when reloading the commands. The bug has been fixed so that whenever we quit the session either by exiting the world or quitting the game directly and launching it back in, the entities don't get summoned a second time.

The two downsides of this are that the boss will reset everytime we exit the world, and that it's a bit less optimized than before (for now, at least)