Water Boss v0.11
The Column, Depth and Row armor stands now have a designated linear trajectory. It has a 50/50 percent change on moving left or right / up or down / forward or backward depending on their axis. They currently bounce on the invisible walls. This allows for a more randomized way to make the boss enter and it feels less scripted to play.