Water Boss v0.12
The boss does not reset when quitting the game by summoning new entities, in fact it keeps the same ones and stocks it somewhere else when the player leaves the world. This allows to keep progress of the boss fight without having to stock the data to a placeholder entity.

The phases are stock inside a scoreboard variable to have more control over the conduct of the fight. However, as of now, the boss duplicates when exiting the game directly. There hasn't been a solution to this unexplicable case so far.