Water Boss v0.13
The datapack now keeps the score data from previous session to newer ones. Meaning that the boss will not reset automatically when exiting the world or the game, unless we manually decide to reset it.

The only downside at the moment is that we cannot reset the fight using the /reload command for now. To do so, we need to add "water_boss:reload" as a string to the load.json file, but it will reset the fight everytime we close the game. This can be used only for debugging.