Water Boss v0.16
The phantom now scatters in one direction and reflects to the next while being able to control the turning speed successfully.

Although, due to how floating-point values are converted as score integers in Minecraft, some working values are not always divisible and are instead approximated.

For instance, while changing the turning speed score to something like 1000, 750, 500, 400, 375, 300, 250, 200, 150 would make the phantom turn only once, non-matching values like 425, 350, 325, 275 would make it turn in place after a revolution.

This is something to be careful about in the eventuality of having to change the rotation speed.

The next step would be to make the movement feel more natural with the use of a sine wave (An armor stand getting teleported in front of a rotating armor stand, storing the relative coordinates in scoreboards for all 3 axis and implementing the score to the position of armor stands)

+ Updated ReadMe